Timeout should be parameter
# problem
5.seconds.fromNow() is too short for slack webSockecet Client.
Usually many server is use 20 second ping/pong setting, so it should be
longer than that. Otherwise it throw SocketConnectionError

Slack send ping pong like 20 second interval, so for websockets, this
should be like 1.minute.fromNow().

# Fix
This PR made deadline (timeout second) to be parameter.

# Result
I’ve tested and 5.seconds throw SocketConnectionError,
60.seconds.fromNow() can keep connection at least 2-3 hours

# Merge-ability
This adds timeout parameter on WebSocket Class but it has default value
so it should not make API conflict or changes. Should be safe to tag it.